backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "bcrypt==4.0.1",
    "chromadb>=1.5.5",
    "cryptography>=46.0.5",
    "deepface>=0.0.99",
    "email-validator>=2.3.0",
    "fastapi>=0.135.1",
    "firebase-admin>=7.2.0",
    "google-auth>=2.49.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.3.0",
    "httpx>=0.28.1",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.11",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.0.10",
    "numpy>=2.4.3",
    "openai>=2.26.0",
    "opencv-python>=4.13.0.92",
    "opencv-python-headless>=4.13.0.92",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.48",
    "tf-keras>=2.20.1",
    "uvicorn>=0.41.0",
]
